Remove legacy add-on communication channel code
- Merge frontend/src/app/lib/moz.js into InstallManager.js since we only
  use mozAddonManager now

- Tweak InstallManager functions to quietly do nothing if
  mozAddonManager is unavailable

- Delete old legacy add-on comms channel in
  frontend/src/app/lib/addon.js

- Also fixes an error with CustomEvent in IE11

Fixes #2744
